+++ crates/store/src/server/api.rs
@@ -2,7 +2,10 @@ use std::{collections::BTreeSet, sync::Arc};
 
 use miden_node_proto::{
     convert,
-    domain::{accounts::AccountInfo, notes::NoteAuthenticationInfo},
+    domain::{
+        accounts::{AccountInfo, AccountProofRequest},
+        notes::NoteAuthenticationInfo,
+    },
     errors::ConversionError,
     generated::{
         self,
@@ -477,25 +480,25 @@ impl api_server::Api for StoreApi {
         &self,
         request: Request<GetAccountProofsRequest>,
     ) -> Result<Response<GetAccountProofsResponse>, Status> {
-        let request = request.into_inner();
-        if request.account_ids.len() < request.code_commitments.len() {
-            return Err(Status::invalid_argument(
-                "The number of code commitments should not exceed the number of requested accounts.",
-            ));
-        }
-
         debug!(target: COMPONENT, ?request);
+        let GetAccountProofsRequest {
+            account_requests,
+            include_headers,
+            code_commitments,
+        } = request.into_inner();
 
-        let include_headers = request.include_headers.unwrap_or_default();
-        let account_ids: Vec<AccountId> = read_account_ids(&request.account_ids)?;
-        let request_code_commitments: BTreeSet<RpoDigest> = try_convert(request.code_commitments)
-            .map_err(|err| {
-            Status::invalid_argument(format!("Invalid code commitment: {}", err))
-        })?;
+        let include_headers = include_headers.unwrap_or_default();
+        let request_code_commitments: BTreeSet<RpoDigest> = try_convert(code_commitments)
+            .map_err(|err| Status::invalid_argument(format!("Invalid code commitment: {}", err)))?;
+
+        let account_requests: Vec<AccountProofRequest> =
+            try_convert(account_requests).map_err(|err| {
+                Status::invalid_argument(format!("Invalid account proofs request: {}", err))
+            })?;
 
         let (block_num, infos) = self
             .state
-            .get_account_proofs(account_ids, request_code_commitments, include_headers)
+            .get_account_proofs(account_requests, request_code_commitments, include_headers)
             .await?;
 
         Ok(Response::new(GetAccountProofsResponse {